fix(sec12): rollback file locations, FK column types, work-order attachment routes
- move 297-299 rollback files from database/migrations/ to database/rollbacks/
  (migrate.js would have applied them as migrations; CI counts and numbering
  checks would both fail)
- ticket_attachments + work_order_attachments: INT UNSIGNED -> BIGINT UNSIGNED
  on id/FK columns (referenced PKs are BIGINT UNSIGNED; MySQL rejects
  mismatched FK types, ER 3780)
- add work-order attachment endpoints (list/upload/delete/download) — the
  table and its 3 permissions were seeded with no consuming routes
- document the 4 new endpoints in openapi.js, regenerate spec (630 paths)
- gitignore uploads/

+++ database/migrations/300_create_ticket_attachments.sql
@@ -1,14 +1,14 @@
 -- Migration 300 — Ticket Attachments (§12)
 CREATE TABLE IF NOT EXISTS ticket_attachments (
-  id               INT UNSIGNED NOT NULL AUTO_INCREMENT,
-  ticket_id        INT UNSIGNED NOT NULL,
+  id               BIGINT UNSIGNED NOT NULL AUTO_INCREMENT,
+  ticket_id        BIGINT UNSIGNED NOT NULL,
   filename         VARCHAR(255) NOT NULL,
   original_filename VARCHAR(255) NOT NULL,
   mime_type        VARCHAR(100) NOT NULL,
   file_size        INT UNSIGNED NOT NULL,
   storage_path     VARCHAR(500) NOT NULL,
-  uploaded_by      INT UNSIGNED NOT NULL,
-  organization_id  INT UNSIGNED NOT NULL,
+  uploaded_by      BIGINT UNSIGNED NOT NULL,
+  organization_id  BIGINT UNSIGNED NOT NULL,
   created_at       TIMESTAMP    NOT NULL DEFAULT CURRENT_TIMESTAMP,
   PRIMARY KEY (id),
   KEY idx_ticket_attachments_ticket (ticket_id),
